refactor(global_lock): move tests to global_lock_tests.rs
Satisfies the pre-push test-file convention that forbids inline
test blocks (tests must live in `*_tests.rs` siblings).
